found bug causing large processing delays so now can sample at roughly 80kHz. Now it appears the issue is more related to too small an input signal

+++ Core/Inc/cirbuf.h
@@ -10,7 +10,7 @@
 
 #include <stdint.h>
 
-float ArrayIdxToCirBufIdx(
+uint8_t ArrayIdxToCirBufIdx(
 		uint8_t start_idx,
 		int8_t array_idx,
 		uint8_t cirbuf_len);
